Improved range selector code
