+++ src/app.ts
@@ -6,7 +6,7 @@ import * as plugins from 'phovea_core/src/plugin';
 import * as d3 from 'd3';
 import * as events from 'phovea_core/src/event';
 import {AppConstants} from './AppConstants';
-import Timeline from './Timeline';
+import Timeline from './timeline/Timeline';
 
 /**
  * Interface for all Views
